fetch vaccination info from firestore
