+++ scripts/.gitignore
@@ -2,3 +2,4 @@
 slack-summary.txt
 county-history-data.json
 vaccination_alerts/vaccination-alerts.json
+google-service-account.json
