+++ scripts/common/firebase/index.ts
@@ -1,0 +1,3 @@
+import { getFirestore } from './firestore';
+
+export { getFirestore };
